refactor(agent): remove model handler type indirection
