+++ src/agent/modelHandlers/google/modelHandlerGoogleInteractions.ts
@@ -29,7 +29,7 @@ import type {
   ExtractResponseResult,
   GoogleToolCall,
   TokenCountOptions,
-} from '@agent/types/IModelHandler';
+} from '@agent/types/ModelHandlerContracts';
 import {
   detectStatusCode,
   attachPartialText,
