+++ src/test-kernel/tools/BashToolErrorFeedback.vitest.ts
@@ -11,7 +11,7 @@ import { AgentWorkspaceState } from '@agent/core/state/AgentWorkspaceState';
 // Local imports - model handlers
 import { ModelHandlerOpenAIResponse } from '@agent/modelHandlers/openai/modelHandlerOpenAIResponse';
 import { extractToolAttachments } from '@agent/core/tools/toolAttachmentExtraction';
-import type { OpenAIResponseToolCall } from '@agent/types/IModelHandler';
+import type { OpenAIResponseToolCall } from '@agent/types/ModelHandlerContracts';
 import { BASH_APPROVAL_CONFIG_KEY } from '@shared/schemas/agentCliSettings';
 
 // Local imports - tools
